Adds loot comparator info popup
Implements a popup to guide users on how to retrieve
chest logs and explains the color codes for item slots
in the loot comparator.

The popup is displayed on mouse enter and hidden on mouse leave
from the info icon next to the chest file upload button.

Also includes some code cleanup such as using collection initializers
and adds support for more DateTime Formats.